changed Sequence and GlyphCodec get methods to take longs instead of ints.  All implementations still use ints to internally store their data (since they are array based) but this will allow us to change that in the future.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1326 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/util/slice/CompactedSliceMap.java
@@ -71,7 +71,7 @@ public final class CompactedSliceMap implements SliceMap {
                 * {@inheritDoc}
                 */
                 @Override
-                public PhredQuality get(int index) {
+                public PhredQuality get(long index) {
                     return defaultQuality;
                 }
